discord_bot/requirements.txt
python-dotenv~=0.20.0
discord~=2.0.0
pydantic~=1.9.2
pytz~=2022.2.1